+++ CHANGES.rst
@@ -8,6 +8,12 @@ Changelog
   jQuery >= 1.7 (Plone 4.3 onwards).
   [gaudenz]
 
+- Register events as Content Rules Event Types if plone.contentrules is present
+  [avoinea]
+
+- Trigger custom events on comment add/remove/reply
+  [avoinea]
+
 - Execute the proper workflow change when using the moderation buttons instead
   of hardcoding the workflow action to always publish
   [omiron]
@@ -19,7 +25,7 @@ Changelog
 - Show author email to Moderator when it is available in anonymous comment.
   [gotcha, smoussiaux]
 
-- Put defaultUser.png instead of old defaultUser.gif 
+- Put defaultUser.png instead of old defaultUser.gif
   [bsuttor]